move terraform tailor under goals
[ci skip-rust]

[ci skip-build-wheels]
