+++ src/python/pants/backend/experimental/terraform/register.py
@@ -2,8 +2,8 @@
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
 from pants.backend.python.util_rules.pex import rules as pex_rules
-from pants.backend.terraform import dependency_inference, style, tailor, target_gen, tool
-from pants.backend.terraform.goals import check
+from pants.backend.terraform import dependency_inference, style, target_gen, tool
+from pants.backend.terraform.goals import check, tailor
 from pants.backend.terraform.lint.tffmt.tffmt import rules as tffmt_rules
 from pants.backend.terraform.target_types import (
     TerraformModulesGeneratorTarget,
